Update README.md
modified the example to include a label I think is very important: lifecycle status. 
Whilst the wording I used is probably not perfect, (retirement, obsolescence, incumbent) a difficulty we have in the industry right now is known when a model should be retired. 

once a model maintainer determines that a model should be retired or replaced, if this was tagged in the ai-bom, then it would make it much easier for ops teams and applied ml teams to know they need to move on also. 

Signed-off-by: adrianw <32903188+5stars217@users.noreply.github.com>